+++ src/ast/validation.cpp
@@ -3,18 +3,17 @@
 #include <stdlib.h>
 
 #include "ast/ast.hh"
-#include "ast/parser.hh"
 #include "ast/validator.hh"
-#include "generated/parser.tab.hh"
+#include "parse/syntax_parser.hh"
 
 void ni::ast::validator::visit(const ni::ast::function_call &node) {
   auto it = functions.find(node.identifier);
   if (it == functions.end()) {
-    throw ni::ast::syntax_error(node.loc, "function \"" + node.identifier +
-                                              "\" not defined");
+    throw ni::parse::syntax_error(node.loc, "function \"" + node.identifier +
+                                                "\" not defined");
   }
   if (node.params.size() != it->second.args.size()) {
-    throw ni::ast::syntax_error(
+    throw ni::parse::syntax_error(
         node.loc, "incorrect number of arguments, function \"" +
                       node.identifier + "\" expects " +
                       std::to_string(it->second.args.size()) + " arguments, " +
@@ -27,8 +26,8 @@ void ni::ast::validator::visit(const ni::ast::function_call &node) {
 void ni::ast::validator::visit(const ni::ast::variable_lookup &node) {
   auto it = vars.find(node.identifier);
   if (it == vars.end()) {
-    throw ni::ast::syntax_error(node.loc, "variable \"" + node.identifier +
-                                              "\" not defined");
+    throw ni::parse::syntax_error(node.loc, "variable \"" + node.identifier +
+                                                "\" not defined");
   }
 
   ni::ast::visitor::visit(node);
@@ -37,8 +36,8 @@ void ni::ast::validator::visit(const ni::ast::variable_lookup &node) {
 void ni::ast::validator::visit(const ni::ast::variable_assignment &node) {
   auto it = this->vars.find(node.identifier);
   if (it == this->vars.end()) {
-    throw ni::ast::syntax_error(node.loc, "variable \"" + node.identifier +
-                                              "\" not defined");
+    throw ni::parse::syntax_error(node.loc, "variable \"" + node.identifier +
+                                                "\" not defined");
   }
   ni::ast::visitor::visit(node);
 }
@@ -46,9 +45,9 @@ void ni::ast::validator::visit(const ni::ast::variable_assignment &node) {
 void ni::ast::validator::visit(const ni::ast::variable_declaration &node) {
   auto it = this->vars.find(node.identifier);
   if (it != this->vars.end()) {
-    throw ni::ast::syntax_error(node.loc, "variable \"" + node.identifier +
-                                              "\" already defined (at " +
-                                              it->second.str() + ")");
+    throw ni::parse::syntax_error(node.loc, "variable \"" + node.identifier +
+                                                "\" already defined (at " +
+                                                it->second.str() + ")");
   }
   this->vars.insert({node.identifier, node.loc});
 
@@ -61,9 +60,9 @@ void ni::ast::validator::visit(const ni::ast::function_declaration &node) {
   for (auto &arg : node.args) {
     auto it = this->vars.find(arg->identifier);
     if (it != this->vars.end()) {
-      throw ni::ast::syntax_error(arg->loc, "argument \"" + arg->identifier +
-                                                "\" already defined (at " +
-                                                it->second.str() + ")");
+      throw ni::parse::syntax_error(arg->loc, "argument \"" + arg->identifier +
+                                                  "\" already defined (at " +
+                                                  it->second.str() + ")");
     }
     this->vars.insert({arg->identifier, arg->loc});
   }
@@ -75,9 +74,9 @@ void ni::ast::validator::visit(const ni::ast::program &node) {
   for (auto &n : node.instructions) {
     auto it = this->functions.find(n->identifier);
     if (it != this->functions.end()) {
-      throw ni::ast::syntax_error(n->loc, "function \"" + n->identifier +
-                                              "\" already defined (at " +
-                                              it->second.loc.str() + ")");
+      throw ni::parse::syntax_error(n->loc, "function \"" + n->identifier +
+                                                "\" already defined (at " +
+                                                it->second.loc.str() + ")");
     }
     this->functions.insert({n->identifier, *n});
   }
